disable lcd timers and manually update the screen. Probably fixes simonedegiacomi/EV3-API#1 simonedegiacomi/EV3-API#6
