--- src/ev3_command.c
@@ -1,33 +1,0 @@
-/*
- * The contents of this file are subject to the Mozilla Public License
- * Version 1.1 (the "License"); you may not use this file except in
- * compliance with the License. You may obtain a copy of the License at
- * http://www.mozilla.org/MPL/
- *
- * Software distributed under the License is distributed on an "AS IS"
- * basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See the
- * License for the specific language governing rights and limitations
- * under the License.
- *
- * The Initial Developer of this code is John Hansen.
- * Portions created by John Hansen are Copyright (C) 2009-2013 John Hansen.
- * All Rights Reserved.
- *
- */
-
-#include "ev3_command.h"
-#include <time.h>
-
-static void ms2ts(struct timespec *ts, unsigned long ms)
-{
-	ts->tv_sec = ms / 1000;
-	ts->tv_nsec = (ms % 1000) * 1000000;
-}
-
-void Wait(unsigned long ms)
-{
-	struct timespec delay;
-	ms2ts(&delay, ms);
-	while (nanosleep(&delay, &delay));
-}
-
