+++ src/icons/large_icons.xbm
