+++ src/icons/normal_icons.xbm
