+++ src/icons/small_icons.xbm
